Extract AdapterTypePicker and AdapterEnvironmentResult for shared use
Pull the adapter-type radio-card grid (recommended + "More" expander) and
the env-test result card out of OnboardingWizard.tsx into their own
components. The wizard now consumes them, and a follow-up Coach onboarding
flow can reuse both without duplicating the visual contract or the env
gating logic.

- New ui/src/components/AdapterTypePicker.tsx — pure presentation; consumer
  owns the model-default side effects via onChange.
- New ui/src/components/AdapterEnvironmentResult.tsx — same component
  previously inlined at the bottom of the wizard.
- OnboardingWizard.tsx drops disabledTypes, the recommended/more memo, the
  showMoreAdapters toggle, the inline AdapterEnvironmentResult definition,
  and the now-unused listUIAdapters / isVisualAdapterChoice / getAdapterDisplay /
  useDisabledAdaptersSync imports. The model-default switch on adapter
  change is preserved verbatim in the new onChange callback.
